+++ app/controllers/settings/two_factor_authentications_controller.rb
@@ -1,57 +1,30 @@
 # frozen_string_literal: true
 
-class Settings::TwoFactorAuthenticationsController < ApplicationController
-  layout 'admin'
+module Settings
+  class TwoFactorAuthenticationsController < ApplicationController
+    layout 'admin'
 
-  before_action :authenticate_user!
-  before_action :verify_otp_required, only: [:new]
+    before_action :authenticate_user!
+    before_action :verify_otp_required, only: [:create]
 
-  def show; end
+    def show; end
 
-  def new
-    current_user.otp_secret = User.generate_otp_secret(32)
-    current_user.save!
-    prepare_two_factor_form
-  end
-
-  def create
-    if current_user.validate_and_consume_otp!(confirmation_params[:code])
-      flash[:notice] = I18n.t('two_factor_authentication.enabled_success')
-
-      current_user.otp_required_for_login = true
-      @recovery_codes = current_user.generate_otp_backup_codes!
+    def create
+      current_user.otp_secret = User.generate_otp_secret(32)
       current_user.save!
-
-      render 'settings/recovery_codes/index'
-    else
-      flash.now[:alert] = I18n.t('two_factor_authentication.wrong_code')
-      prepare_two_factor_form
-      render :new
+      redirect_to new_settings_two_factor_authentication_confirmation_path
     end
-  end
 
-  def destroy
-    current_user.otp_required_for_login = false
-    current_user.save!
-
-    redirect_to settings_two_factor_authentication_path
-  end
-
-  private
-
-  def verify_otp_required
-    if current_user.otp_required_for_login?
+    def destroy
+      current_user.otp_required_for_login = false
+      current_user.save!
       redirect_to settings_two_factor_authentication_path
     end
-  end
 
-  def prepare_two_factor_form
-    @confirmation = Form::TwoFactorConfirmation.new
-    @provision_url = current_user.otp_provisioning_uri(current_user.email, issuer: Rails.configuration.x.local_domain)
-    @qrcode = RQRCode::QRCode.new(@provision_url)
-  end
+    private
 
-  def confirmation_params
-    params.require(:form_two_factor_confirmation).permit(:code)
+    def verify_otp_required
+      redirect_to settings_two_factor_authentication_path if current_user.otp_required_for_login?
+    end
   end
 end
