+++ app/views/settings/two_factor_authentication/recovery_codes/index.html.haml
