+++ app/views/settings/two_factor_authentications/show.html.haml
@@ -2,16 +2,25 @@
   = t('settings.two_factor_authentication')
 
 .simple_form
-  %p.hint= t('two_factor_authentication.description_html')
+  %p.hint
+    = t('two_factor_authentication.description_html')
 
   - if current_user.otp_required_for_login
-    = link_to t('two_factor_authentication.disable'), settings_two_factor_authentication_path, data: { method: :delete }, class: 'block-button'
+    = link_to t('two_factor_authentication.disable'),
+      settings_two_factor_authentication_path,
+      data: { method: :delete },
+      class: 'block-button'
   - else
-    = link_to t('two_factor_authentication.setup'), new_settings_two_factor_authentication_path, class: 'block-button'
+    = link_to t('two_factor_authentication.setup'),
+      settings_two_factor_authentication_path,
+      data: { method: :post },
+      class: 'block-button'
 
 - if current_user.otp_required_for_login
-  %p
-
   .simple_form
-    %p.hint= t('two_factor_authentication.lost_recovery_codes')
-    = link_to t('two_factor_authentication.generate_recovery_codes'), settings_two_factor_authentication_recovery_codes_path, data: { method: 'POST' }, class: 'block-button'
+    %p.hint
+      = t('two_factor_authentication.lost_recovery_codes')
+    = link_to t('two_factor_authentication.generate_recovery_codes'),
+      settings_two_factor_authentication_recovery_codes_path,
+      data: { method: :post },
+      class: 'block-button'
