+++ config/routes.rb
@@ -58,8 +58,10 @@ Rails.application.routes.draw do
       resources :mutes, only: :index, controller: :muted_accounts
     end
 
-    resource :two_factor_authentication, only: [:show, :new, :create, :destroy] do
+    resource :two_factor_authentication, only: [:show, :create, :destroy]
+    namespace :two_factor_authentication do
       resources :recovery_codes, only: [:create]
+      resource :confirmation, only: [:new, :create]
     end
   end
