+++ spec/controllers/settings/two_factor_authentication/recovery_codes_controller_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe Settings::RecoveryCodesController do
+describe Settings::TwoFactorAuthentication::RecoveryCodesController do
   render_views
 
   let(:user) { Fabricate(:user) }
